PaddleFSL/requirements.txt
paddlepaddle-gpu==2.4.1
rdkit-pypi
pgl
paddlehelix
paddlefsl
scikit-learn==1.0.2
ujson
paddlenlp==2.0.6
transformers==3.3.1
Pillow==8.2.0